Added test for unstaging selected lines. Added the unstage test and moved the staging tests to a new file.
